requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Deep Learning & NLP
torch>=1.10.0
transformers>=4.18.0
sentence-transformers>=2.2.0
nltk>=3.6.0
faiss-cpu>=1.7.0  # Use faiss-gpu if using GPU

# API Integration
openai>=0.27.0

# Utilities
python-dotenv>=0.19.0
diskcache>=5.2.0
streamlit>=1.10.0

# Testing
pytest>=6.0.0

# Logging
logging>=0.4.9